move selectTracing to accessor file
